### 2025-03-15 #7898 @editorialbot
[PRE REVIEW]: fastfrechet: An R package for fast implementation of Fréchet regression with distributional responses
**Submitting author:** @alexandercoulter (<a href="http://orcid.org/0009-0001-9116-6588">Alexander Coulter</a>)
**Repository:** https://github.com/alexandercoulter/fastfrechet
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
