### 2025-03-17 #7898 @fabian-s
@alexandercoulter  please check & fix the DOIs found by the bot while I try to rustle up some reviewers:

```
❌ MISSING DOIs

- 10.32614/cran.package.wri may be a valid DOI for title: WRI: Wasserstein Regression and Inference
- 10.32614/cran.package.frechet may be a valid DOI for title: frechet: Statistical Analysis for Random Objects a...
- 10.32614/cran.package.microbenchmark may be a valid DOI for title: microbenchmark: Accurate Timing Functions
```
