### 2025-03-17 #7898 @fabian-s
@jdtuck thank you!  I'll provide more instructions on the review process for JOSS once I've found a 2nd reviewer..
