### 2025-03-17 29b73ca
Update REFERENCES to include bot-recommended DOIs for CRAN package citations.
