+++ inst/REFERENCES.bib
@@ -235,6 +235,7 @@
 	url = {https://CRAN.R-project.org/package=WRI},
 	author = {Liu, Xi and Zhang, Chao and Coleman, Matthew and Petersen, Alexander},
 	year = {2022},
+	doi = {10.32614/cran.package.wri},
 	note = {R package version 0.2.0},
 }
 
@@ -243,6 +244,7 @@
 	url = {https://CRAN.R-project.org/package=frechet},
 	author = {Chen, Yaqing and Zhou, Yidong and Chen, Han and Gajardo, Alvaro and Fan, Jianing and Zhong, Qixian and Dubey, Paromita and Han, Kyunghee and Bhattacharjee, Satarupa and Alexander, Petersen and Müller, Hans-Georg},
 	year = {2023},
+	doi = {10.32614/cran.package.frechet},
 	note = {R package version 0.3.0},
 }
 
@@ -251,5 +253,6 @@
 	url = {https://CRAN.R-project.org/package=microbenchmark},
 	author = {Mersmann, Olaf},
 	year = {2024},
+	doi = {10.32614/cran.package.microbenchmark},
 	note = {R package version 1.5.0},
 }
