### 2025-03-19 alexandercoulter/fastfrechet#1 @jdtuck
Add contributions policies
Please update `README.md` on how you would like people to provide updates to the package from the community. In other words how do you want someone to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support. Thank you

openjournals/joss-reviews#7925
