### 2025-03-19 #7925 @jdtuck
The compiling of the vignettes when building the R package take a long time (probably longer than CRAN will allow). You might want to make that optional when you deploy to CRAN (which you should)
