### 2025-03-19 #7925 @alexandercoulter
I will consider adding the vignettes as articles, this may be a good middle ground.
