### 2025-03-19 #7925 @fabian-s
> Reconstructing vignettes (at least particularly violative ones) as fully pre-made html files which don't need to be recreated e.g. during vignette build will entirely remove the 3.5.0 dependency requirements.

based on my CRAN submission experience, pre-compiled vignettes are really the only way to include substantial examples w/o violating their build time rules.   i'd recommend that if you set up your CI/testing so that it regularly re-compiles them to ensure the documentation stays relevant/correct as the package evolves.
