+++ README.Rmd
@@ -38,3 +38,11 @@ install.packages("devtools")
 devtools::install_github("https://github.com/alexandercoulter/fastfrechet", build_vignettes = TRUE)
 ```
 
+## Community Guidelines
+
+
+1.  Contributions and suggestions to the software are always welcome.
+    Please consult our [contribution guidelines](.github/CONTRIBUTING.md) prior
+    to submitting a pull request.
+2.  Report issues or problems with the software using github’s [issue tracker](https://github.com/alexandercoulter/fastfrechet/issues).
+3.  Contributors must adhere to the [Code of Conduct](CODE_OF_CONDUCT.md).
