### 2025-03-19 alexandercoulter/fastfrechet#1 @irinagain
Thank you for catching this! The readme is now updated with links to contribution guidelines, code of conduct, and issue tracker.
