+++ vignettes/intro-fastfrechet.Rmd
@@ -8,7 +8,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r, include = FALSE}
+```{r knitr-options, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
@@ -53,7 +53,7 @@ The `generate_zinbinom_qf` function from `fastfrechet` generates covariate-depen
 
 `generate_zinbinom_qf` also optionally takes "baseline" parameters of the `zinbinom` distribution, namely, the zero-inflation factor, and the probability and size parameters of the `nbinom` component. As output, `generate_zinbinom_qf` returns a list containing the `n` $\times$ `p` covariate matrix `X`; and the `n` $\times$ `m` response matrix `Y`, which stores the quantile functions row-wise.
 
-```{r}
+```{r generate-data}
 n <- 200 # number of samples - nrow(X) and nrow(Y).
 p <- 20 # number of covariates - ncol(X).
 m <- 200 # quantile function grid density - ncol(Y).
@@ -75,7 +75,7 @@ Y <- mydata$Y # (n x m) matrix of quantile functions, stored row-wise
 
 We illustrate these quantile function responses with line plots. Note the true functions are piece-wise constant, where the line plots interpolate between points on the observed `m`-grid.
 
-```{r}
+```{r plot-EQFs}
 matplot(mseq, t(Y),
   xlim = c(0, 1), ylim = range(Y), type = "l", lty = 1, col = 1,
   main = "znbinom Quantile Functions", xlab = "p", ylab = "quantile"
@@ -104,7 +104,7 @@ The `frechetreg_univar2wass` function from `fastfrechet` takes as input
 
 Since the `zinbinom` distribution is bounded from below by zero, we set `lower = 0`, and use default settings for other parameters. As output, `frechetreg_univar2wass` returns a list containing the set of fitted quantile functions `Qhat`, stored row-wise as in `Y`; and the `n` $\times$ `m+1` matrix of Lagrange multipliers associated with the solution, also stored row-wise.
 
-```{r}
+```{r run-frechetreg}
 lower <- 0
 upper <- Inf
 # Estimate conditional quantile functions:
@@ -130,7 +130,7 @@ min(output$Qhat)
 
 The Fréchet means associated with the Fréchet regression problem $\widehat{\mathbf{Q}}$ are projections of the unconstrained Euclidean response object $\widehat{\mathbf{Y}}$ onto the space of quantile functions. Row-wise, the matrix $\widehat{\mathbf{Y}}$ generally does not obey quantile function constraints of monotonicity or box constraints, whereas the solution $\widehat{\mathbf{Q}}$ does. We recover the unconstrained matrix `Yhat` from the output of `frechetreg_univar2wass`, and illustrate it side-by-side the fitted $\widehat{\mathbf{Q}}$.
 
-```{r, fig.width = 8, fig.height = 5.5}
+```{r plot-unconstrained-constrained, fig.width = 8, fig.height = 5.5}
 # Plot the unconstrained estimators and the constrained conditional quantile functions:
 Yhat <- output$Qhat - output$Lagrange_Multiplier[, -(m + 1)] + output$Lagrange_Multiplier[, -1]
 par(mfrow = c(1, 2), las = 1)
@@ -169,7 +169,7 @@ The `FRiSO_univar2wass` function from `fastfrechet` takes as input
 
[…]
