+++ vignettes/performanceExample-fastfrechet.Rmd
@@ -161,7 +161,7 @@ solving procedure 15 times for each method. We can see the implementation from
 `fastfrechet` is the fastest of the four methods, upward of $1000\times$ faster
 than methods currently available in R packages.
 
-```{r, eval = FALSE}
+```{r microbenchmark-frechetreg, eval = FALSE}
 # Benchmark times over 5 iterations:
 M1 = microbenchmark("WRI" = wass_regress( ~ ., X_df, "quantile", Y_adj),
                     "frechet" = GloDenReg(X, qin = Y, optns = list(lower = 0)),
@@ -175,7 +175,7 @@ M1$time = round(M1$time / 1e6, 1) * 1e6
 M1
 ```
 
-```{r, echo = FALSE}
+```{r show-frechetreg-times, echo = FALSE}
 # Load times:
 s = system.file("private", "microbenchmark_results.Rda", package = "fastfrechet")
 if(file.exists(s)) load(s)
@@ -197,7 +197,7 @@ implementation from `frechet` violates the lower box constraint despite its
 explicit inclusion as an argument. The implementation from the Supplementary
 Material of @tucker_variable_2023 and the implementation from `fastfrechet` both
 solve the problem to numerical accuracy.
-```{r}
+```{r fit-QFs}
 # Fit quantile functions:
 WRI_Q = wass_regress( ~ ., X_df, "quantile", Y_adj)$Qfit
 frechet_Q = GloDenReg(X, qin = Y, optns = list(lower = 0))$qout
@@ -417,8 +417,8 @@ M2$time = round(M2$time / 1e6, 1) * 1e6
 M2
 ```
 
-```{r, echo = FALSE}
-# Round frechetreg times:
+```{r show-FRiSO-times, echo = FALSE}
+# Round FRiSO times:
 M2_temp = M2
 M2_temp$time = round(M2_temp$time / 1e6, 1) * 1e6
 
@@ -439,7 +439,7 @@ solutions. As expected, the `fastfrechet` solutions become more accurate as the
 error tolerance parameter decreases, with only modest increases in computation
 time.
 
-```{r calculate-lambdas}
+```{r obtain-FRiSO-lambdas}
 # Set sequence of error tolerances:
 eseq = 0.014 / c(1, 2, 4, 8)
 
