### 2025-03-19 #7925 @jdtuck
> If that is the route we pursue, then we'll hold off updating the DESCRIPTION file with the implicit R v. 3.5.0 dependency that is being created at the vignette build stage. Reconstructing vignettes (at least particularly violative ones) as fully pre-made html files which don't need to be recreated e.g. during vignette build will entirely remove the 3.5.0 dependency requirements.

Still having a minimum R version in the `DESCRIPTION` is a good idea
