### 2025-03-19 a04a6e0
Add explicit R version dependency (>= 3.5.0).
