+++ DESCRIPTION
@@ -10,6 +10,8 @@ Authors@R:
            comment = c(ORCID = "0000-0002-4116-0268")))
 Description: Fast implementation of Fréchet regression associated functionalities for univariate distribution responses equipped with 2-Wasserstein metric, including weighted Fréchet means and Fréchet Ridge Selection Operator (FRiSO).
 License: MIT + file LICENSE
+VignetteBuilder: knitr
+Depends: R (>= 3.5.0)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
@@ -28,5 +30,4 @@ Suggests:
     WRI,
     frechet,
     microbenchmark
-VignetteBuilder: knitr
 Config/testthat/edition: 3
