### 2025-03-19 47e04b0
Update intro vignette to reference pre-saved results, to reduce compute time at build.
