### 2025-03-19 61aae1b
Modify performanceExample vignette to reference pre-saved results to reduce compute time at build.
