### 2025-03-19 #7925 @alexandercoulter
> Still having a minimum R version in the `DESCRIPTION` is a good idea

That's fair.  Then I've added pre-saved results into the inst/private folder which the vignettes reference, in order to reduce computation time.  I've decided in light of your recommendation to keep the .Rda file dependency, and simply require dependence on version 3.5.0 or later.  This is now included in the DESCRIPTION file.

Total vignette computation time is now approximately 3.6 seconds on my end.  I hope this resolves both issues @jdtuck!
