### 2025-03-19 df3d321
Include vignettes/*.png files in gitignore file.
