### 2025-04-07 alexandercoulter/fastfrechet#3 @Almond-S
More specific critique of other approaches
Please be more precise and careful in your critique of other approaches in the Statement of Need. In particular, here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L61
(which is obviously too general - probably you have the UK Biobank data in the back of your mind)
but also here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L67

(in light of openjournals/joss-reviews#7925)
