### 2025-04-07 alexandercoulter/fastfrechet#4 @Almond-S
Advantage of auxiliary weighting scheme
Here it would be great if you could very briefly indicate the advantage of this novel feature:
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L78-L80
It seems relevant but not self-explanatory.

(in context of openjournals/joss-reviews#7925)
