### 2025-04-07 alexandercoulter/fastfrechet#5 @Almond-S
Indicate ground truth for simulation studies
I think it would be good if you could at least very briefly indicate how the data is generated, here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L86
and/or at least here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/performanceExample-fastfrechet.Rmd#L79
without details, and potentially give a reference for details.

(in light of openjournals/joss-reviews#7925)
