### 2025-04-07 alexandercoulter/fastfrechet#6 @Almond-S
Small formulation thingi
Sorry, this is a bit miniscule but "discussed in" seems a bit much here:
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L127-L128
Maybe "mentionned in" or "referred to"?

(in context of openjournals/joss-reviews#7925)
