### 2025-04-07 #7925 @alexandercoulter
@Almond-S Thank you for your notes, I will try to accommodate some of the requests for more explanation while maintaining the desired word count limit.  @fabian-s, while the bot reads over 1000 words in the paper.md file, the rendered paper by my rough count was somewhere closer to 950 (right up to the start of References).  Could you please clarify how final word count is judged?
