### 2025-04-07 alexandercoulter/fastfrechet#7 @Almond-S
Constraint comparison figure
I think the constraint comparison figure, shown here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/paper.md?plain=1#L150
and here
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/performanceExample-fastfrechet.Rmd#L317
could easily carry more information than it does right now, avoiding over-plotting:
- choosing a subsample of the curves would allow identification of single examples (perhaps in colors)
- in grey in the back you might consider either showing stoll all predicted curves, or alternatively always the fastfrechet predictions for close comparison
- certainly it makes sense to focus on the constrain issue. But the selected regions seem very small, while it might be interesting to see a bit more of the curves to assess how close the predictions of the different methos are overall (if they are very close, couldn't one also simply set everything below zero simply post-hoc to zero, as a quite reasonable work-around for the alternative approaches -- while yours is of course still much faster?)

(in context of openjournals/joss-reviews#7925)
