### 2025-04-07 alexandercoulter/fastfrechet#8 @Almond-S
low-rank solutions
Apologies, I might be very slow, but what do you mean by the design matrix X being "low-rank" here:
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/performanceExample-fastfrechet.Rmd#L132
? It seems this is particularly challenging since
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/performanceExample-fastfrechet.Rmd#L143-L144
but also that X is not singular (at least in your examples).  Or is that actually the point and you can handle the singular case due to the regularization introduced?

(in context of openjournals/joss-reviews#7925)
