### 2025-04-07 alexandercoulter/fastfrechet#9 @Almond-S
prediction example
I think it would be very nice if you could add a small example using
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/intro-fastfrechet.Rmd#L100
for prediction at covariates outside of the training data to the vignette. I know that this seems sort of trivial, but I was first looking for a prediction function and missed the Z argument somehow. So, I believe, as this is so often needed, it would be very kind to point your users more explicitly to it.

(in context of openjournals/joss-reviews#7925)
