### 2025-04-07 alexandercoulter/fastfrechet#10 @Almond-S
Copy paste typo
Probably there is a little typo here:

https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/intro-fastfrechet.Rmd#L298

and `FRiSO_CV_univar2wass` should be `FRiSO_CPSS_univar2wass`.

(in context of openjournals/joss-reviews#7925)
