### 2025-04-07 alexandercoulter/fastfrechet#11 @Almond-S
Definition of \hat{y}
I believe it would be good to introduce the \hat{y} here briefly:
https://github.com/alexandercoulter/fastfrechet/blob/df3d321496411bd0ddd421c5eceec275465a35e4/vignettes/monotoneQP-fastfrechet.Rmd#L23
They do appear already in the introductory vignette but in any case a comment would be useful - and since they re-appear, perhaps also an a bit more detailed introduction than now would be good here or there.

(In context of openjournals/joss-reviews#7925)
