### 2025-04-10 alexandercoulter/fastfrechet#6 @alexandercoulter
Hi @Almond-S, thank you for raising these issues.  I've got time to begin addressing them and will start with the shorter ones like this.  I agree that "discussed in []" is overstating things, and I think "mentioned in []" would be appropriate.  I will make that change now.
