### 2025-04-10 2b6b805
Correct copy-paste error where FRiSO_CV... function was named in a paragraph instead of FRiSO_CPSS...
