### 2025-04-11 1dc3ad9
Add more specificity to speed comparisons to other packages and methods, with explicit reference to speed demonstrations in subsequent paragraphs and other referenced work(s).
