### 2025-04-11 alexandercoulter/fastfrechet#3 @alexandercoulter
I've included more specific references to time and speed comparisons in the earlier Coulter et al. (2024) paper and the subsequent paragraphs where we directly run the competing method on simulation settings.  For instance, for the "prohibitively slow" statement regarding the algorithm from Tucker et al. (2023), I included specific data size and run-time from Coulter et al. (2024) to illustrate that computation was slow even on modestly sized data.  Hopefully this is less vague/general!
