### 2025-04-11 48c0046
Update Fréchet regression comparison figure creation code in vignette, and remove some extra white space in paper.md.
