+++ vignettes/performanceExample-fastfrechet.Rmd
@@ -191,12 +191,13 @@ M1_temp
 ## Optimization Accuracy Comparison
 
 We compare the four methods on optimization accuracy for the Fréchet regression
-on the example data set. We can see the implementation from `WRI` does not obey
-known box constraints because the option to specify them is not available; the
-implementation from `frechet` violates the lower box constraint despite its
-explicit inclusion as an argument. The implementation from the Supplementary
-Material of @tucker_variable_2023 and the implementation from `fastfrechet` both
-solve the problem to numerical accuracy.
+on the example data set, highlighting two example curves by red and blue color.
+We can see the implementation from `WRI` does not obey known box constraints
+because the option to specify them is not available; the implementation from
+`frechet` violates the lower box constraint despite its explicit inclusion as an
+argument. The implementation from the Supplementary Material of
+@tucker_variable_2023 and the implementation from `fastfrechet` both solve the
+problem to numerical accuracy.
 ```{r fit-QFs, eval = FALSE}
 # Fit quantile functions:
 WRI_Q = wass_regress( ~ ., X_df, "quantile", Y_adj)$Qfit
@@ -212,7 +213,11 @@ if(file.exists(s)) load(s)
 ```
 
 ```{r figure2, include = FALSE}
-# Plot quantile functions, focusing around zero:
+# Pick specific lines to plot:
+pick = c(54, 69)
+plotcols = c('#648FFF', '#785EF0', '#DC267F', '#FFB000', 'black')
+
+# Figure margin and size specifications:
 # 0.4 + 4a + 4(0.06) = 6.5
 a = (6.5 - 0.32 - 0.45) / 4
 mais = cbind(c(0.35, 0.35, 0.35, 0.35),
@@ -224,6 +229,7 @@ figs = cbind(c(0, 0.4 + a + 0.04, 0.4 + 2 * (a + 0.04), 0.4 + 3 * (a + 0.04)) / 
              c(0, 0, 0, 0),
              c(1, 1, 1, 1))
 
+# Plot quantile functions, focusing around zero:
 png("frechetreg_accuracy_comparison.png", width = 6.5, height = 2.2, units = "in", res = 576)
 
 par(fig = figs[1, ], mai = mais[1, ], family = "Helvetica")
@@ -234,6 +240,7 @@ polygon(x = par("usr")[c(1, 2, 2, 1)], y = par("usr")[c(3, 3, 4, 4)],
 abline(h = axTicks(2), v = axTicks(1), col = "white")
 abline(h = 0, lty = 2, col = "gray50")
 matlines(mseq, t(WRI_Q), col = "black", lty = 1, lwd = 1.5)
+matlines(mseq, t(WRI_Q[pick, ]), col = plotcols[c(1, 3)], lty = 1, lwd = 2)
 axis(1, axTicks(1), c("0", "0.2", "0.4", "0.6", "0.8", 1), FALSE, -1.3, cex.axis = 0.55)
 axis(2, NULL, TRUE, FALSE, -0.9, las = 1, cex.axis = 0.55)
 mtext("u", 1, 0.7, cex = 0.8)
@@ -257,6 +264,7 @@ polygon(x = par("usr")[c(1, 2, 2, 1)], y = par("usr")[c(3, 3, 4, 4)],
 abline(h = axTicks(2), v = axTicks(1), col = "white")
[…]
