### 2025-04-12 10f9009
Add language and new figure for Fréchet regression comparison, imposing constraints post hoc.
