### 2025-04-12 465149c
Add language to manuscript indicating that even applying support constraints after the fact, the solutions from WRI and frechet remain sub-optimal.
