### 2025-04-12 alexandercoulter/fastfrechet#7 @alexandercoulter
Hi @Almond-S , I've made some adjustments in light of your comments.

1. We believe the first point you bring up is a good suggestion, and newer images include selected curves in Red and Blue to help track how the different methods fit them.

2. We were concerned that adding more curves into the background would create very busy graphs, especially since they are already "spaghetti-like" graphs.  I hope the current version with the Red/Blue curves helps illustrate sufficiently.

3. Since the support boundaries are where violations happen most obviously, we've chosen to remain focused on those regions.  I believe it is true that imposing support constraints _post hoc_ should give an identical solution as otherwise including them explicitly, however it does not seem fruitful to prove this.  However, to demonstrate what applying constraints after the fitting procedure does for overall fit, especially of the WRI and frechet methods, we have included a new graphic in the performanceExample vignette.  It illustrates the difference in fitted objective function values of the WRI, frechet, and Tucker methods v. our fastfrechet method, respectively.  We illustrate that even after applying support constraints _post hoc_, the WRI and frechet solutions are sub-optimal compared to fastfrechet.  The differences can be up to order 10^-3.  The Tucker method is nearly identical, with differences on the order of 10^-13 or smaller.

We do not explicitly include this latter graphic in the revised manuscript, but we reference the conclusion drawn from it and direct the reader to look within the performanceExample vignette for more details.

Have a look and let us know whether these changes are satisfactory :)
