### 2025-04-14 alexandercoulter/fastfrechet#7 @Almond-S
Great, thanks!
I think this helps a lot to get a feeling of what the differences are.
