### 2025-04-14 alexandercoulter/fastfrechet#3 @Almond-S
Thank you for including more details here, which already helps a lot.
However, I still consider the general terminus "prohibitively" as too pejorative and would recommend to replace or mitigate it - for instance, with something like "is quite slow", "is relatively slow", "can likely be problematically slow", "might be prohibitively slow in many applications, for example ..." etc. (Unless I am not aware that it is commonly used as a specific terminus in this context.)
By contrast, you describe the other two approaches as "comparatively slow" even though they are - if I didn't get it wrong - much slower than the Tucker approach.
