### 2025-04-14 014fe1a
Remove "prohibitively slow" language and replace with more specific "infeasible" language WRT resampling methods like CPSS.
