### 2025-04-14 alexandercoulter/fastfrechet#3 @alexandercoulter
We appreciate the concern about pejorative language and the apparent discrepancy to language used elsewhere.  I have removed the language about it being "prohibitively" slow, and instead have opted to be more specific to the original motivation—making resampling methods feasible—for the new variable selection algorithm.

The algorithm is now described as "can be slow in many applications, for example..." (adopting your language), and the stronger language about "infeasible" is now applied more specifically with respect to tasks which are estimated to take several CPU-days to complete with the old algorithm.
