### 2025-04-14 c2a5bf6
Update return value to correct output, closer to the monotoneQP function which this function mimics.
