### 2025-04-14 0177fe6
Add specific definition of the "m-seq" used for the distribution support, as named specifically in the vignettes etc.
