### 2025-04-14 aba5a5f
Add more explicit reference to the performanceExample vignette for description on how example zinbinom quantile functions are generated.
