### 2025-04-14 fa3c298
Detail the QF generating scheme implemented in the "generate_zinbinom_qf" function.
