+++ vignettes/performanceExample-fastfrechet.Rmd
@@ -75,12 +75,43 @@ $\pmb{Y} \in \mathbb{R}^{n\times m}$, where $m$ is the density of an equidistant
 grid in $(0, 1)$ on which each quantile function is evaluated.
 
 The data we utilize for our comparisons are simulated covariate-dependent
-zero-inflated negative binomial distributions. The `fastfrechet` function
-`generate_zinbinom_qf` allows us to simulate $n$ such distributions, represented
-as quantile functions evaluated on a shared $m$-grid in $(0, 1)$, dependent on
-the first 4 of $p \geq 4$ covariates. The example data set we generate for this
-vignette has sample size $n = 100$, covariate number $p = 10$, and quantile
-functions evaluated on an equidistant $m = 100$ grid.
+zero-inflated negative binomial (`zinbinom`) distributions, which are
+parameterized by a zero-inflation parameter $z$, size parameter $r$, and
+probability parameter $\rho$. The `fastfrechet` function `generate_zinbinom_qf`
+allows us to simulate $n$ such distributions, represented as quantile functions
+evaluated on a shared $m$-grid in $(0, 1)$, dependent on the first 4 of
+$p \geq 4$ covariates. Specifically, we generate covariates by
+$$
+\mathbf{x}_i \sim \mathcal{N}(\pmb{0}, \mathbf{I}_p), \qquad \forall i = 1, \dots, n.
+$$
+Then for each $i$, we generate a quantile function by first sampling the
+distribution parameters via the following equations (dropping the $i$ subscript
+for convenience)
+$$
+\begin{align*}
+\mathrm{logit}(z) &\sim \mathcal{N}(\alpha_z + \beta_z \cdot x_4 \:, \: s_z^2), \\
+\mathrm{log}(r) &\sim \mathcal{N}(\alpha_r + \beta_r \cdot (x_2 + x_3) \:, \: s_r^2), \\
+\mathrm{logit}(\rho) &\sim \mathcal{N}(\alpha_{\rho} + \beta_{\rho} \cdot x_1 \:,\: s_{\rho}^2).
+\end{align*}
+$$
+We use default parameters
+$$
+\begin{align*}
+\begin{aligned}
+\alpha_z &= \mathrm{logit}(0.2), &\qquad \beta_z &= 0.4, &\qquad s_z^2 &= 0.1^2,\\
+\alpha_r &= \log(10), &\qquad \beta_r &= 0.2, &\qquad s_r^2 &= 0.15^2, \\
+\alpha_{\rho} &= \mathrm{logit}(0.5) &\qquad \beta_{\rho} &= 0.2, &\qquad s_{\rho}^2 &= 0.15^2.
+\end{aligned}
+\end{align*}
+$$
+Finally, the quantile functions are generated by
+$$
+\mathbf{q}_{\mathbf{x}_i}( u \mid z_i, r_i, \rho_i) = F^{-1}_{\mathrm{nbinom}} \left( \frac{ u - z_i}{1 - z_i} \mid r_i, \rho_i  \right), \qquad u \in \left\{ \begin{matrix} \frac{1}{2m} & \frac{3}{2m} & \cdots & \frac{2m-1}{2m} \end{matrix} \right\},
+$$
+where $F^{-1}_{\mathrm{nbinom}}$ is effectuated by calling the `qnbinom`
+function in `R`. The example data set we generate for this vignette has sample
+size $n = 100$, covariate number $p = 10$, and quantile functions evaluated on
+an equidistant $m = 100$ grid.
 
 ```{r generate-ZINB}
 set.seed(31)
