### 2025-04-14 7fd2429
Update documentation files for functions.
