### 2025-04-14 alexandercoulter/fastfrechet#5 @alexandercoulter
Since a more verbose description in the vignette would be helpful for the reader, I have added more language detailing how the QF's (and covariates) are generated by the included function.  The paper now also makes a more explicit reference for the reader, pointing them to the performanceExample vignette for details on the generation steps.
