### 2025-04-14 1fe1e85
additional details on auxillary weighting scheme needed for variable selection
