### 2025-04-14 0e92d36
additional details on auxillary weighting scheme needed for variable selection
