+++ paper.md
@@ -75,15 +75,11 @@ is comparatively slow (i.e. takes upward of 10,000$\times$ longer than
 The `fastfrechet` package addresses these limitations by providing a fast,
 scalable, and user-friendly implementation of both Fréchet regression and
 variable selection for 2-Wasserstein space, based on the work of
-@coulter_fast_2024. The package incorporates resampling tools to enhance
+@coulter_fast_2024. The Fréchet regression solver features a customized dual active-set algorithm, inspired by @arnstrom_dual_2022, which ensures both
+computational efficiency and accuracy while accommodating user-specified support
+constraints. To support variable selection, it is also the first Fréchet regression solver to incorporate an auxiliary weighting scheme. In this scheme, the covariate-dependent weights that determine each observation's influence can be modified using a user-supplied vector $\lambda$ [@tucker_variable_2023], which specifies which covariates are excluded from the weight construction. The package incorporates resampling tools to enhance
 automatic variable selection, including cross-validation described in
 @tucker_variable_2023 and stability selection described in @coulter_fast_2024.
-Additionally, `fastfrechet` features a customized dual active-set solver for the
-Fréchet regression problem, inspired by @arnstrom_dual_2022, which ensures both
-computational efficiency and accuracy while accommodating user-specified support
-constraints. The Fréchet regression solver also accommodates the auxiliary
-weighting scheme used in the variable selection procedure, the first package to
-do so.
 
 # Performance Comparisons to Existing Implementations
 
