### 2025-04-14 alexandercoulter/fastfrechet#4 @irinagain
Thank you for catching this. We have added the clarifications to make it self-contained and slightly rearranged the paragraph for logical flow. Specific to your question, it now reads:

 To support variable selection, it is also the first Fréchet regression solver to incorporate an auxiliary weighting scheme. In this scheme, the covariate-dependent weights that determine each observation's influence can be modified using a user-supplied vector λ [@tucker_variable_2023], which specifies which covariates are excluded from the weight construction.
