### 2025-04-18 e50fa88
(1) Clarify requirements on Z in the case X is not full column rank; (2) call new QR-decomposition based functions to perform fast calculation of Yhat in the event lambda is not specified; (3) implicitly implement row-space check for (1 Z), including new exit condition if the rows of (1 Z) are not in the row-space of (1 X); (4) include new parameter dictating the tolerance of the row-space check; (5) add checks for new parameter value (must be strictly positive scalar); (6) re-format file.
