+++ R/frechetreg_univar2wass.R
@@ -10,37 +10,58 @@
 #' set of covariate vectors. Accommodates the regularization scheme from
 #' \insertCite{tucker_variable_2023}{fastfrechet}.
 #'
-#' @param X A (`n` \eqn{\times} `p`) "input" covariate matrix with no missing, all finite entries.
+#' @param X A (`n` \eqn{\times} `p`) "input" covariate matrix with no missing, all finite entries. We recommend you **do not** include an intercept column (i.e. all `1`'s vector).
 #' @param Y A (`n` \eqn{\times} `m`) matrix of observed quantile functions evaluated on a shared uniform `m`-grid in \eqn{(0, 1)} - e.g. `seq(0.5/m, 1 - 0.5/m, len = m)` - row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by `lower` and `upper` parameters.
-#' @param Z An optional (`z` \eqn{\times} `p`) "output" covariate matrix (default `NULL`) with no missing, all finite entries.
+#' @param Z An optional (`z` \eqn{\times} `p`) "output" covariate matrix (default `NULL`) with no missing, all finite entries. If the rows of \eqn{(\begin{matrix} 1 & \mathbf{Z} \end{matrix})} do not belong to the column space of \eqn{(\begin{matrix} 1 & \mathbf{X} \end{matrix})^{\top}}, then there is no unique Fréchet regression solution.
 #' @param C_init An optional (`n` \eqn{\times} `m + 1`) matrix (or `z` \eqn{\times} `m + 1` if `Z` is provided) of non-negative entries, specifying initial active set(s) for optimization. Active sets are identified by positive entries, row-wise if a matrix.
 #' @param lambda An optional (`p` \eqn{\times} `1`) vector with non-negative entries whose sum is strictly positive.
 #'  If `NULL` a non-regularized regression is performed (the default).
 #' @param lower An optional numeric scalar (default `-Inf`) lower box constraint; must be strictly less than `upper`.
 #' @param upper An optional numeric scalar (default `Inf`) upper box constraint; must be strictly greater than `lower`.
 #' @param eps An optional numeric scalar (default `1e-10`) error tolerance; must be strictly positive.
+#' @param Ztol An optional numeric scalar (default `1e-10`) error tolerance for the row-space condition on `Z`; must be strictly positive. See Details.
 #'
 #' @details
-#' Fréchet regression generalizes Euclidean regression to the general metric space setting.
-#' In the sample setting, we observe covariate-response pairs \eqn{\{(\mathbf{x}_i, \mathbf{y}_i)\} \subset \mathbb{R}^p \times \Omega},
-#' where \eqn{\mathbf{x}_i}'s are rows of matrix \eqn{\mathbf{X}} column-centered
-#' and scaled such that \eqn{\mathrm{diag}(\mathbf{X}^{\top}\mathbf{X}) = \pmb{1}},
-#' and \eqn{\Omega} is the space of univariate quantile functions equipped with
-#' the 2-Wasserstein metric \eqn{d_W(\mathbf{q}, \mathbf{p}) = \lVert\mathbf{q} - \mathbf{p}\rVert_{L_2[0,1]}}.
-#' (Generally we also only observe the distributions over a discrete grid, not as
-#' full functions.) The conditional Fréchet mean is estimated by
[…]
