### 2025-04-18 807a1cd
Create new C++ based functions for performing QR decomposition, and for performing projection step
Yhat = (1 Z)((1 X)'(1 X))^{-1}(1 X)'Y.

Functions include check for whether Z obeys the requisite row-space constraints.
