### 2025-04-18 4daafda
Update exports file to include two new Rcpp functions.
