### 2025-04-18 2febce6
Remove comment about post-hoc bound application.
