### 2025-04-18 2bfc94d
Update exports file with new Rcpp functions.
