### 2025-04-18 a92142a
Add new tests to ensure frechetreg fails under certain conditions where Z matrix does not satisfy affine combination condition.
