### 2025-04-18 9bef16b
Update function documentation, especially for frechetreg where new conditions on Z are clarified.
