+++ man/frechetreg_univar2wass.Rd
@@ -12,15 +12,16 @@ frechetreg_univar2wass(
   lambda = NULL,
   lower = -Inf,
   upper = Inf,
-  eps = 1e-10
+  eps = 1e-10,
+  Ztol = 1e-06
 )
 }
 \arguments{
-\item{X}{A (\code{n} \eqn{\times} \code{p}) "input" covariate matrix with no missing, all finite entries.}
+\item{X}{A (\code{n} \eqn{\times} \code{p}) "input" covariate matrix with no missing, all finite entries. We recommend you \strong{do not} include an intercept column (i.e. all \code{1}'s vector).}
 
 \item{Y}{A (\code{n} \eqn{\times} \code{m}) matrix of observed quantile functions evaluated on a shared uniform \code{m}-grid in \eqn{(0, 1)} - e.g. \code{seq(0.5/m, 1 - 0.5/m, len = m)} - row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}
 
-\item{Z}{An optional (\code{z} \eqn{\times} \code{p}) "output" covariate matrix (default \code{NULL}) with no missing, all finite entries.}
+\item{Z}{An optional (\code{z} \eqn{\times} \code{p}) "output" covariate matrix (default \code{NULL}) with no missing, all finite entries. If the rows of \eqn{(\begin{matrix} 1 & \mathbf{Z} \end{matrix})} do not belong to the column space of \eqn{(\begin{matrix} 1 & \mathbf{X} \end{matrix})^{\top}}, then there is no unique Fréchet regression solution.}
 
 \item{C_init}{An optional (\code{n} \eqn{\times} \code{m + 1}) matrix (or \code{z} \eqn{\times} \code{m + 1} if \code{Z} is provided) of non-negative entries, specifying initial active set(s) for optimization. Active sets are identified by positive entries, row-wise if a matrix.}
 
@@ -32,6 +33,8 @@ If \code{NULL} a non-regularized regression is performed (the default).}
 \item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}
 
 \item{eps}{An optional numeric scalar (default \code{1e-10}) error tolerance; must be strictly positive.}
+
+\item{Ztol}{An optional numeric scalar (default \code{1e-10}) error tolerance for the row-space condition on \code{Z}; must be strictly positive. See Details.}
 }
 \value{
 A list object with components:
@@ -51,26 +54,46 @@ set of covariate vectors. Accommodates the regularization scheme from
 \insertCite{tucker_variable_2023}{fastfrechet}.
 }
 \details{
-Fréchet regression generalizes Euclidean regression to the general metric space setting.
-In the sample setting, we observe covariate-response pairs \eqn{\{(\mathbf{x}_i, \mathbf{y}_i)\} \subset \mathbb{R}^p \times \Omega},
-where \eqn{\mathbf{x}_i}'s are rows of matrix \eqn{\mathbf{X}} column-centered
-and scaled such that \eqn{\mathrm{diag}(\mathbf{X}^{\top}\mathbf{X}) = \pmb{1}},
-and \eqn{\Omega} is the space of univariate quantile functions equipped with
-the 2-Wasserstein metric \eqn{d_W(\mathbf{q}, \mathbf{p}) = \lVert\mathbf{q} - \mathbf{p}\rVert_{L_2[0,1]}}.
-(Generally we also only observe the distributions over a discrete grid, not as
[…]
