+++ vignettes/performanceExample-fastfrechet.Rmd
@@ -25,14 +25,14 @@ library(microbenchmark)
 ```
 
 ```{r load-code, include = FALSE}
-s = system.file("private", "GlobalObj4h.R", package = "fastfrechet")
-if(file.exists(s)) source(s)
-s = system.file("private", "GloWassReg.R", package = "fastfrechet")
-if(file.exists(s)) source(s)
-s = system.file("private", "IndivRidgeGloWassReg.R", package = "fastfrechet")
-if(file.exists(s)) source(s)
-s = system.file("private", "lambdaGlobalcoordesc.R", package = "fastfrechet")
-if(file.exists(s)) source(s)
+s <- system.file("private", "GlobalObj4h.R", package = "fastfrechet")
+if (file.exists(s)) source(s)
+s <- system.file("private", "GloWassReg.R", package = "fastfrechet")
+if (file.exists(s)) source(s)
+s <- system.file("private", "IndivRidgeGloWassReg.R", package = "fastfrechet")
+if (file.exists(s)) source(s)
+s <- system.file("private", "lambdaGlobalcoordesc.R", package = "fastfrechet")
+if (file.exists(s)) source(s)
 ```
 
 # Purpose of Vignette
@@ -109,22 +109,22 @@ $$
 \mathbf{q}_{\mathbf{x}_i}( u \mid z_i, r_i, \rho_i) = F^{-1}_{\mathrm{nbinom}} \left( \frac{ u - z_i}{1 - z_i} \mid r_i, \rho_i  \right), \qquad u \in \left\{ \begin{matrix} \frac{1}{2m} & \frac{3}{2m} & \cdots & \frac{2m-1}{2m} \end{matrix} \right\},
 $$
 where $F^{-1}_{\mathrm{nbinom}}$ is effectuated by calling the `qnbinom`
-function in `R`. The example data set we generate for this vignette has sample
+function in R. The example data set we generate for this vignette has sample
 size $n = 100$, covariate number $p = 10$, and quantile functions evaluated on
 an equidistant $m = 100$ grid.
 
 ```{r generate-ZINB}
 set.seed(31)
-n = 100        # number of quantile functions
-p = 10         # number of covariates
-m = 100        # (0, 1) quantile functions grid density 
+n <- 100 # number of quantile functions
+p <- 10 # number of covariates
+m <- 100 # (0, 1) quantile functions grid density
 
 # Equidistant m-grid in (0, 1)
-mseq = seq(0.5 / m, 1 - 0.5 / m, length.out = m)
+mseq <- seq(0.5 / m, 1 - 0.5 / m, length.out = m)
 
-gendata = generate_zinbinom_qf(n, p, m)
-X = gendata$X  # (n x p) covariate matrix
-Y = gendata$Y  # (n x m) quantile response matrix, stored row-wise
+gendata <- generate_zinbinom_qf(n, p, m)
+X <- gendata$X # (n x p) covariate matrix
+Y <- gendata$Y # (n x m) quantile response matrix, stored row-wise
 ```
 
 ```{r figure1, include = FALSE}
[…]
