### 2025-04-18 27676f2
Begin process of including low-rank settings; update formatting with "styler" package.
