### 2025-04-18 27d928d
Clean up paragraph right-indenting; add extensive discussion of low-rank setting, including conditions on Z; illustrate how if the rows of Z are not affine combinations of the rows of X while X is not full column rank, then the frechetreg function throws an error.
