+++ vignettes/intro-fastfrechet.Rmd
@@ -26,32 +26,74 @@ library(fastfrechet)
 
 # Introduction
 
-`fastfrechet` is an R package providing fast and robust solutions to several computational problems associated with Fréchet regression [@petersen_frechet_2019] and an associated variable selection method [@tucker_variable_2023], focusing on a specific metric space setting. Fréchet regression generalizes regression with Euclidean covariates and Euclidean responses, to settings where responses exist in a general metric space (i.e. non-vector space). The space of univariate distribution response objects is one such example, recently gaining popularity in biomedical applications where technological advances have made rich patient-level data sets commonplace in a variety of settings. These larger data sets and complex response constraints require newer, faster algorithms for feasibility and scalability [@coulter_fast_2024], which `fastfrechet` provides.\
+`fastfrechet` is an R package providing fast and robust solutions to several
+computational problems associated with Fréchet regression
+[@petersen_frechet_2019] and an associated variable selection method
+[@tucker_variable_2023], focusing on a specific metric space setting. Fréchet
+regression generalizes regression with Euclidean covariates and Euclidean
+responses, to settings where responses exist in a general metric space (i.e.
+non-vector space). The space of univariate distribution response objects is one
+such example, recently gaining popularity in biomedical applications where
+technological advances have made rich patient-level data sets commonplace in a
+variety of settings. These larger data sets and complex response constraints
+require newer, faster algorithms for feasibility and scalability
+[@coulter_fast_2024], which `fastfrechet` provides.\
 
-This vignette walks through an example use case of all of the available functions in `fastfrechet`. This includes simulating example covariate-dependent distribution responses, implementing Fréchet regression, performing variable selection, and utilizing resampling tools to supplement automatic variable selection. Where appropriate, the mathematical problems being solved by each function are briefly described. A second vignette, `monotoneQP-fastfrechet`, provides a more detailed algorithmic exposition of the function `monotoneQP` specifically.
+This vignette walks through example use cases of all of the available functions
+in `fastfrechet`. This includes simulating example covariate-dependent
+distribution responses, implementing Fréchet regression, performing variable
+selection, and utilizing resampling tools to supplement automatic variable
+selection. Where appropriate, the mathematical problems being solved by each
+function are briefly described. A second vignette, `monotoneQP-fastfrechet`,
+provides a more detailed algorithmic exposition of the function `monotoneQP`
+specifically.
 
 
 # Running Through Examples
 
[…]
