### 2025-04-18 6e2f299
Clarify that WRI cannot take a Z input, and frechet/Tucker implementations can; reduce graphics size to keep package size small.
