### 2025-04-18 2e9484d
Include mention of Z fuctionality (or lack thereof) in different packages; direct readers to the intro vignette in order to see more details on the low-rank and Z functionality in the fastfrechet::frechetreg function.
