### 2025-04-18 b54bc5e
Remove redundant graph of Q(|Z) for affine example.
