### 2025-04-18 alexandercoulter/fastfrechet#8 @alexandercoulter
Thank you for your questions about this, especially since our vignettes did not elaborate on this further.  Since it is a more complicated setting, for which users need some more instruction, we took our time to ensure our checks, math, and results were robust for the settings in which the Fréchet regression problem does, or does not, have a solution.

I'll begin with a summary of our changes, and then discuss some of the background in comment here.

1. We have added new C++ functions for calculating the unconstrained "Yhat" object in the absence of a specified "lambda" input, which perform QR decomposition and check a specific row-space condition on input Z, if it is provided.  (More below.)
2. The documentation of the `frechetreg_univar2wass` function has been updated to instruct users on what conditions are placed on Z when X is not full column rank.
3. We have added new tests to ensure the `frechetreg_univar2wass` function handles the Z matrix correctly.
4. The `intro-fastfrechet` vignette has been updated with details on the low-rank setting, as well as some demonstrations of the `frechetreg_univar2wass` handling of the low-rank case.  This includes when Z is not specified, and the function's error handling when Z is specified.
5. The `performanceExample-fastfrechet` vignette has been given a minor update to direct the user to the `intro-fastfrechet` vignette for more details on the low-rank setting.  While you have quoted from the `performanceExample` vignette, we believe the `intro` vignette is a more appropriate place to discuss this issue.

I also wanted to give a little more explanation to facilitate answering your questions.  The "low rank" setting, i.e. when X is not full column rank, is difficult because the unconstrained solution for the Fréchet regression problem—absent any regularization from `lambda`—can be characterized

Yhat | Z = (1 Z)G(1 X)'Y.

Here, G is any generalized inverse of (1 X)'(1 X).  When X is full column rank, there is a unique G (the simple inverse), and the solution follows as written.  When X is not full column rank, there are infinitely many generalized inverses of (1 X)'(1 X).  To maintain estimability, we require the rows of (1 Z) to belong to the row-space of (1 X).  While this immediately holds when Z = X, it is not generally true for arbitrary Z.  We elaborate on this a little more in the `intro` vignette.

Your questions then, and our responses:

> (1) It seems this is particularly challenging since [excerpt][]. (2) Or is that actually the point and you can handle the singular case due to the regularization introduced?

1. The row-space condition on Z is easily checkable via e.g. the QR-decomposition methods we implement.  This is perhaps a little more challenging than trying to invert X'X, as the previous implementations do.
2. We do not rely on regularization via `lambda` (or otherwise), but instead characterize the conditions under which Yhat | Z is linearly estimable.
