### 2025-04-18 26656c6
Remove explicit return of "Yhat" object.
