### 2025-04-22 5edd19d
Include two examples where the optional Z matrix is specified.
