### 2025-04-22 alexandercoulter/fastfrechet#9 @llrebecca21
Thank you for this suggestion. I have included two examples where the optional matrix Z is specified. The first example shows what happens when a covariate known to be related to the responses changes value. The second example demonstrates what occurs when a covariate known to be unrelated to the responses changes value.
