### 2025-04-23 alexandercoulter/fastfrechet#8 @Almond-S
Thank you for clarifying this so thoroughly. Another strength of your package comes out much more clearly now.
