### 2025-04-23 alexandercoulter/fastfrechet#9 @Almond-S
Great, I think such a nice illustration is very helpful.
