### 2025-04-23 alexandercoulter/fastfrechet#4 @Almond-S
Ah great, thanks, then this is clear now.
