### 2025-04-23 alexandercoulter/fastfrechet#11 @Almond-S
Maybe this was not so clear or sounds more work-intense than it should:
you do already introduce the $\hat{Y}$ and $\hat{Q}$ in plain words in the intro vignette and plot them for the example. I think simply making the introducting more explicit, stating how they arise from what you define in
https://github.com/alexandercoulter/fastfrechet/blob/5edd19d3502644d345add474d2c74b0bd48479cd/vignettes/intro-fastfrechet.Rmd#L144
(or perhaps a modified version for $\hat{Y}$?), plus an explicit reference from the respective paragraph of the monotoneQP vignette to the intro vignette, would be absolutely sufficient.
It's only that right now one has a fealing of it, but its definition as "unconstrained Euclidean response object $\widehat{\mathbf{Y}}$" doesn't give me full certainty that I truly know what you are talking about here.
