### 2025-04-23 8e05012
Clarify the "y-hat" equations in the montoneQP vignette, and point readers to the intro-vignette for details on the constraints on x (and implications for y).
