+++ vignettes/monotoneQP-fastfrechet.Rmd
@@ -17,13 +17,21 @@ knitr::opts_chunk$set(
 
 ## Background
 
-This vignette details the mathematical foundation of the algorithm we implement to solve the non-convex Fréchet mean problem for univariate distribution responses equipped with the 2-Wasserstein metric. In the "population setting", this is an orthogonal projection problem. In the "sample setting" where distribution responses are represented by quantile functions $\mathbf{y} \in \mathbb{R}^m$ evaluated on a discrete uniform $m$-grid in $(0, 1)$ and obeying box constraints $y_j \in [b_L, b_U] \subseteq \mathbb{R}$, the Fréchet mean problem can be written
+This vignette details the mathematical foundation of the algorithm we implement
+to solve the non-convex Fréchet mean problem for univariate distribution
+responses equipped with the 2-Wasserstein metric. In the "population setting",
+this is an orthogonal projection problem. In the "sample setting" where
+distribution responses are represented by quantile functions
+$\mathbf{y} \in \mathbb{R}^m$ evaluated on a discrete uniform $m$-grid in
+$(0, 1)$ and obeying box constraints $y_j \in [b_L, b_U] \subseteq \mathbb{R}$,
+the Fréchet mean problem can be written
 
 $$
 \widehat{\mathbf{q}} := \mathrm{arg\:min}_{\mathbf{q} \in \mathbb{R}^m} \lVert{\mathbf{q} - \widehat{\mathbf{y}}} \rVert_2^2, \qquad \mathbf{b} - \mathbf{A}^{\top}\mathbf{q} \leq \pmb{0}_{m+1},
 $$
 where $\mathbf{b}^{\top} = \begin{pmatrix} b_L & \pmb{0}_{m-1}^{\top} & -b_U\end{pmatrix} \in \mathbb{R}^{m+1}$,
-$\mathbf{A} \in \mathbb{R}^{m\times(m+1)}$ is a first-difference matrix of the form
+$\mathbf{A} \in \mathbb{R}^{m\times(m+1)}$ is a first-difference matrix of the
+form
 
 $$
 \mathbf{A} = \begin{pmatrix}
@@ -36,12 +44,27 @@ $$
    0 &  0 &  0 &  0 & \cdots &  0 & +1 & -1
 \end{pmatrix},
 $$
-and finally $\widehat{\mathbf{y}}$ is a non-convex sum of the $\mathbf{y}$'s
-which does not necessarily sit in the target space. The final projected vector $\widehat{\mathbf{q}}$ is monotone non-decreasing (i.e. $q_i \leq q_j$ for $i \leq j$) and obeys the same box constraints as the $\mathbf{y}$'s. The `fastfrechet` package implements a solver for this problem via the `monotoneQP` function.
+Finally, in Fréchet regression, $\widehat{\mathbf{y}}$ is a non-convex sum of
+the observed response $\mathbf{y}$'s which does not necessarily sit in the
+target space, defined for some $\mathbf{x}_* \in \mathbb{R}^p$
+$$
+\widehat{\mathbf{y}}^{\top} = (n^{-1}\pmb{1}^{\top} + \mathbf{x}_*^{\top}(\mathbf{X}^{\top}\mathbf{X})^-\mathbf{X}^{\top})\mathbf{Y}.
+$$
+so long as $\mathbf{x}_*$ obeys a suitable row-space condition. (See the
+accompanying `intro-fastfrechet` vignette for details.) This is a quadratic
+programming (QP) problem, where the final projected vector
+$\widehat{\mathbf{q}}$ is monotone non-decreasing (i.e. $q_i \leq q_j$ for
+$i \leq j$) and obeys the same box constraints as the $\mathbf{y}$'s. While the
[…]
