### 2025-04-23 alexandercoulter/fastfrechet#11 @alexandercoulter
Hi @Almond-S, my apologies for the delays in responding to this.  It seemed appropriate to push this one off until the low-rank setting one was finished, since some of the important details in the `intro-vignette` relating to X and Y—and hence Y-hat—were in flux at that time.  But I'm glad those changes are satisfactory!

For this issue, I agree the previous wording was a little unclear, and have changed it to emphasize two points:

1. In direct response to your comment, I've included the explicit equation for a specific y-hat (though the solution to the argmin problem you've since quoted, not the argmin characterization itself).  The `monotoneQP-vignette` also directs users to the `intro-vignette` to get more details.
2. I've clarified a distinction that was elided in the original wording, that `frechetreg_univar2wass` solves the Fréchet regression problem while `monotoneQP` solves the more general QP problem of finding a monotone+bounded vector/matrix that is "closest" to the input object.
