### 2025-04-23 d528e46
Split paragraphs and clarify what "This [is a quadratic programming problem]" is referring to : qhat, not yhat.
