### 2025-04-23 alexandercoulter/fastfrechet#11 @Almond-S
Okay, great, in the `monotoneQP-vignette` it is clear now, thank you!
Only, in the `intro-vignette` I am afraid it is not entirely clear, yet.
Maybe here
https://github.com/alexandercoulter/fastfrechet/blob/d528e46a1fb6e720e63bf25e294af72078f63747/vignettes/intro-fastfrechet.Rmd#L195-L196
just something like
``
$\widehat{\mathbf{Q}}$ are projections onto the space of quantile functions of corresponding predictions $\widehat{\mathbf{Y}}$ as they would be obtained by Fréchet regression without monotonicity and box constraints.
``
or something alike. I think not even formulae are needed (eventhough they could maybe even make it more clear). Just a little bit more explicit definition.
