### 2025-04-23 67e02c5
Update language and equations clarifying what yhat is, and how qhat more directly relates to it.
