+++ vignettes/intro-fastfrechet.Rmd
@@ -135,17 +135,28 @@ $\{(\mathbf{x}_i, \mathbf{y}_i)\} \subset \mathbb{R}^p \times \Omega$, where
 $\Omega$ is the space of univariate quantile functions equipped with the
 2-Wasserstein metric
 $d_W(\mathbf{q}, \mathbf{p}) = \lVert\mathbf{q} - \mathbf{p}\rVert_{L_2[0,1]}$.
-(Generally we also only observe the distributions over a discrete grid, not as
-full functions.) We treat the covariate vectors $\mathbf{x}_i$ as rows of matrix
+Generally we also only observe the distributions over a discrete grid, not as
+full functions, hence the 2-Wasserstein metric reduces to a finite-dimension
+Euclidean norm. We treat the covariate vectors $\mathbf{x}_i$ as rows of matrix
 $\mathbf{X}$, which is column-centered and scaled so that
-$\mathrm{diag}(\mathbf{X}^{\top}\mathbf{X}) = \pmb{1}$. The conditional Fréchet
-mean at $X = \mathbf{x}_*$ is estimated by
+$\mathrm{diag}(\mathbf{X}^{\top}\mathbf{X}) = \pmb{1}$.
+
+The conditional Fréchet mean at $X = \mathbf{x}_*$ is estimated by
 $$
-\widehat{\mathbf{q}}_{\oplus}(\mathbf{x}_*) := \underset{\mathbf{q}\in\Omega}{\mathrm{argmin}} \sum_{i=1}^n \left\{ \frac{1}{n} + \mathbf{x}_*^{\top}(\mathbf{X}^{\top}\mathbf{X})^+\mathbf{x}_i\right\} d_W^2(\mathbf{y}_i, \mathbf{q}).
+\begin{align}
+\widehat{\mathbf{q}}_{\oplus}(\mathbf{x}_*) &:= \underset{\mathbf{q}\in\Omega}{\mathrm{argmin}} \sum_{i=1}^n \left\{ \frac{1}{n} + \mathbf{x}_*^{\top}(\mathbf{X}^{\top}\mathbf{X})^-\mathbf{x}_i\right\} d_W^2(\mathbf{y}_i, \mathbf{q}), \\
+\Rightarrow \qquad \widehat{\mathbf{q}}_{\oplus}(\mathbf{x}_*) &= \underset{\mathbf{q}\in\Omega}{\mathrm{argmin}} \lVert \mathbf{q} - \widehat{\mathbf{y}}(\mathbf{x}_*) \rVert_2^2, \\
+\mbox{where} \qquad \widehat{\mathbf{y}}(\mathbf{x}_*) &= (n^{-1}\pmb{1}^{\top} + \mathbf{x}_*^{\top}(\mathbf{X}^{\top}\mathbf{X})^- \mathbf{X}^{\top}) \mathbf{Y}.
+\end{align}
 $$
-This is the Fréchet regression problem, a weighted Fréchet mean problem solvable
-by quadratic programming. This is implemented by `frechetreg_univar2wass`, using
-a dual active-set method inspired from @arnstrom_dual_2022.
+The object $\widehat{\mathbf{y}}(\mathbf{x}_*)$ is an unconstrained solution, a
+projection onto the column space of $\mathbf{X}$; the object
+$\widehat{\mathbf{q}}(\mathbf{x}_*)$ is a fitted quantile function, a projection
+of the unconstrained solution onto the constraint space, to obey monotonicity
+and possible support bounds. This is the Fréchet regression problem, a weighted
+Fréchet mean problem solvable by quadratic programming. This is implemented by
+`frechetreg_univar2wass`, using a dual active-set method inspired from
+@arnstrom_dual_2022.
 
 ### Using `fastfrechet::frechetreg_univar2wass`
 
[…]
