### 2025-04-23 alexandercoulter/fastfrechet#11 @alexandercoulter
Ah, I understand where you're coming from.  I've updated language in the `intro-fastfrechet` vignette and added some equations, just to have all of the bases covered.  Hopefully that clarifies in each place what the "yhat" object(s) is/are.
