### 2025-04-23 alexandercoulter/fastfrechet#11 @Almond-S
Very nice, now it's perfectly clear! Thanks
