### 2025-04-23 #7925 @Almond-S
Green light from my side!

Thanks for addressing all issues so thoroughly! Checklist complete.
