### 2025-04-29 #7925 @alexandercoulter
Hi @fabian-s, I have completed these steps.  The archived version's version tag is v1.1.0, and the Zenodo archival DOI is

10.5281/zenodo.15302750
