### 2025-04-29 #7925 @alexandercoulter
I've also ensured the title, authors, version tag, and license match between the Zenodo deposit metadata and the archived package's information from DESCRIPTION/paper.md/LICENSE/etc.
